+++ ccramic/utils/pixel_level_utils.py
@@ -1,6 +1,7 @@
 import dash
 import numpy as np
 import pandas as pd
+import scipy.stats
 from PIL import Image
 from PIL import ImageColor
 import plotly.graph_objects as go
@@ -543,6 +544,7 @@ class MarkerCorrelation:
                                baseline_threshold: Union[float, int]=0, mask: Union[np.array, np.ndarray, None]=None,
                                blend_dict: dict=None, use_blend_params: bool=True, bounds: Union[str, dict]=None):
 
+        self.basic_correlation = None
         # Initialize the intermediate and final outputs to None
         # overlap between baseline and target in mask
         self.baseline_proportion_in_mask = None
@@ -557,28 +559,35 @@ class MarkerCorrelation:
         # the ratio of the target to the baseline inside the mask
         self.target_proportion_relative = None
 
-        if image_dict is not None and roi_selection in image_dict and target_channel in \
-                image_dict[roi_selection] and mask is not None:
+        if image_dict is not None and roi_selection in image_dict and target_channel in image_dict[roi_selection]:
             # TODO: add in validation parse for checking the ROI images to the mask dimensions
             self.image_dict = image_dict
             self.roi_selection = roi_selection
             self.target_threshold = target_threshold
             self.target_channel = target_channel
             self.bounds = self.compute_channel_bounds_from_zoom(bounds)
-            self.mask = mask[np.ix_(range(int(self.bounds[2]), int(self.bounds[3]), 1),
-                                    range(int(self.bounds[0]), int(self.bounds[1]), 1))] if self.bounds else mask
+            self.mask = mask
+            self.baseline_array = None
+            self.baseline_threshold = baseline_threshold
+            if self.bounds and self.mask is not None:
+                self.mask = mask[np.ix_(range(int(self.bounds[2]), int(self.bounds[3]), 1),
+                                    range(int(self.bounds[0]), int(self.bounds[1]), 1))]
             try:
                 self.target_array, self.target_threshold = self.set_target_array_from_blend(image_dict, use_blend_params,
-                                            blend_dict, target_channel, roi_selection, self.bounds)
-                self.set_target_proportion_in_mask()
-                self.baseline_array = None
-                self.baseline_threshold = baseline_threshold
+                                                        blend_dict, target_channel, roi_selection, self.bounds)
+                if self.mask is not None:
+                    self.set_target_proportion_in_mask()
+            except (ValueError, KeyError):
+                pass
+            try:
                 if baseline_channel and baseline_channel in image_dict[roi_selection]:
                     self.baseline_array, self.baseline_threshold = self.set_baseline_array_from_blend(image_dict,
-                                    use_blend_params, blend_dict, baseline_channel, roi_selection, self.bounds)
-                    self.set_baseline_proportion_in_mask()
-                    self.compute_correlation_statistics()
-            except ValueError:
+                                        use_blend_params, blend_dict, baseline_channel, roi_selection, self.bounds)
+                    self.compute_basic_pearson_correlation()
+                    if self.mask is not None:
+                        self.set_baseline_proportion_in_mask()
+                        self.compute_correlation_statistics()
+            except (ValueError, KeyError):
                 pass
 
     @staticmethod
@@ -596,7 +605,8 @@ class MarkerCorrelation:
         Return the proportion of the target channel that is inside the mask, the
         target overlap with the baseline channel inside the mask, and baseline proportion in mask
         """
-        return self.target_proportion_in_mask, self.target_proportion_relative, self.baseline_proportion_in_mask
+        return self.target_proportion_in_mask, self.target_proportion_relative, self.baseline_proportion_in_mask, \
+            self.basic_correlation
 
     def set_target_proportion_in_mask(self):
         """
@@ -626,12 +636,16 @@ class MarkerCorrelation:
         """
         self.marker_overlap_in_mask = np.logical_and(self.target_threshold_in_mask,
                                                 self.baseline_array > float(self.baseline_threshold))
-        target_image = self.image_dict[self.roi_selection][self.target_channel]
-        target_image = target_image[np.ix_(range(int(self.bounds[2]), int(self.bounds[3]), 1),
-                          range(int(self.bounds[0]), int(self.bounds[1]), 1))] if \
-            self.bounds else target_image
-        self.target_proportion_relative = np.sum(target_image[self.marker_overlap_in_mask]) / \
-                    np.sum(target_image[self.target_threshold_in_mask])
+        # target_image = self.image_dict[self.roi_selection][self.target_channel]
+        # target_image = target_image[np.ix_(range(int(self.bounds[2]), int(self.bounds[3]), 1),
+        #                   range(int(self.bounds[0]), int(self.bounds[1]), 1))] if \
+        #     self.bounds else target_image
+        self.target_proportion_relative = np.sum(self.target_array[self.marker_overlap_in_mask]) / \
+                    np.sum(self.target_array[self.target_threshold_in_mask])
+
+    def compute_basic_pearson_correlation(self):
+        self.basic_correlation = float(scipy.stats.pearsonr(
+            self.target_array.flatten(), self.baseline_array.flatten())[0])
 
     @staticmethod
     def set_target_array_from_blend(image_dict, use_blend_params, blend_dict, target_channel, roi_selection,
@@ -652,6 +666,7 @@ class MarkerCorrelation:
             target_threshold = 0
         target_array = target_array[np.ix_(range(int(bounds[2]), int(bounds[3]), 1),
                           range(int(bounds[0]), int(bounds[1]), 1))] if bounds else target_array
+        target_array = np.where(target_array < target_threshold, 0, target_array)
         return target_array, target_threshold
 
     @staticmethod
@@ -673,4 +688,5 @@ class MarkerCorrelation:
             baseline_threshold = 0
         baseline_array = baseline_array[np.ix_(range(int(bounds[2]), int(bounds[3]), 1),
                           range(int(bounds[0]), int(bounds[1]), 1))] if bounds else baseline_array
+        baseline_array = np.where(baseline_array < baseline_threshold, 0, baseline_array)
         return baseline_array, baseline_threshold