+++ tests/test_pixel_level_inputs.py
@@ -233,17 +233,18 @@ def test_window_viewport_settings():
            {'width': '60.0vh', 'height': '30.0vh'}
 
 def test_generate_marker_correlation_information():
-    children = marker_correlation_children(None, None)
+    children = marker_correlation_children(None, None, None, None)
     assert not children
-    children = marker_correlation_children(0.50, 1.00, 1.00)
+    children = marker_correlation_children(0.50, 1.00, 1.00, 0.5)
     span_counts = 0
     for child in children:
         if isinstance(child, html.Span):
             span_counts += 1
     # assert one span for each of the values
-    assert span_counts == 3
+    assert span_counts == 4
     assert children
-    assert len(children) > 4
+    assert len(children) > 5
+
 
 def test_blank_reset_histogram():
     blank_hist = reset_pixel_histogram(True)